test(router): constrain the FIFO test with a build tag instead of a runtime skip
windows-latest failed to vet:

  vet.exe: internal\router\irregular_file_test.go:37:20: undefined: syscall.Mkfifo

The test had a `if runtime.GOOS == "windows" { t.Skip(...) }` guard, which is the wrong
tool: a runtime check cannot save a reference the compiler must still resolve.
syscall.Mkfifo does not exist on Windows, so the package fails to build there long
before any skip could execute.

Moved to irregular_file_unix_test.go behind `//go:build !windows && !plan9`. The
behavior under test is not Unix-specific — the mode check rejects Windows device paths
through the same os.FileMode bits — only the cheapest way to CREATE a non-regular file
is. The portable half (directories, and regular files under device-like paths) stays in
irregular_file_test.go and runs everywhere.

Why local gates missed it: I ran `GOOS=windows go build ./...`, which does not compile
_test.go files, and ran `go vet` only for the host platform. `GOOS=windows go vet ./...`
reproduces the failure exactly and is now what I check. Verified clean for darwin, linux
and windows; the FIFO test still runs on this host (not silently excluded by the tag);
61 packages pass.
